# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SAHOOL Platform API",
    "description": "Agricultural Intelligence Platform - منصة الذكاء الزراعي",
    "version": "15.3.0",
    "contact": {
      "name": "SAHOOL Support",
      "email": "support@sahool.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8095",
      "description": "SAHOOL Agro Advisor - Disease diagnosis, nutrient assessment, and fertilizer planning for Yemen agriculture"
    },
    {
      "url": "http://localhost:8000",
      "description": "SAHOOL Agent Registry Service - Agent card to register"
    },
    {
      "url": "http://localhost:8095",
      "description": "SAHOOL Agro Advisor - Disease diagnosis, nutrient assessment, and fertilizer planning for Yemen agriculture"
    },
    {
      "url": "http://localhost:8000",
      "description": "AI Advisor Service - User question"
    },
    {
      "url": "http://localhost:8113",
      "description": "SAHOOL Alert Service - معرف التنبيه"
    },
    {
      "url": "http://localhost:8111",
      "description": "SAHOOL Astronomical Calendar Service - مفتاح المرحلة"
    },
    {
      "url": "http://localhost:8089",
      "description": "SAHOOL Billing Core | خدمة الفوترة - Complete billing, subscription, and payment management for SAHOOL platform"
    },
    {
      "url": "http://localhost:8100",
      "description": "SAHOOL Crop Health Service - Normalized Difference Vegetation Index"
    },
    {
      "url": "http://localhost:8095",
...
  "paths": {
...
    "/crops": {
      "get": {
        "summary": "List all crops grouped by category",
        "description": "",
        "tags": ["agro-advisor"],
        "x-summary-ar": "",
        "x-description-ar": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/v1/crops": {
      "get": {
        "summary": "قائمة المحاصيل المدعومة للتنبؤ",
        "description": "",
        "tags": ["yield-engine"],
        "x-summary-ar": "قائمة المحاصيل المدعومة للتنبؤ",
        "x-description-ar": "",
        "parameters": [
          {
            "name": "crop_type",
            "in": "path",
            "required": true,
            "description": "",
            "schema": {
              "type": "croptype"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/v1/phenology/crops": {
      "get": {
        "summary": "List all supported crops for phenology detection",
        "description": "قائمة جميع المحاصيل المدعومة لكشف مراحل النمو",
        "tags": ["vegetation-analysis-service"],
        "x-summary-ar": "قائمة جميع المحاصيل المدعومة لكشف مراحل النمو",
        "x-description-ar": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/v1/predict": {
      "post": {
        "summary": "Predict Yield",
        "description": "",
        "tags": ["yield-engine"],
        "x-summary-ar": "",
        "x-description-ar": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service communication"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}